Move amulet editor plugin
