+++ src/builtin_plugins/plugin/amulet/editor/_resources/lang/en.lang
