+++ src/builtin_plugins/plugin/amulet/editor/_signal.py
