+++ src/builtin_plugins/plugin/amulet/editor/plugin.json
@@ -1,5 +1,6 @@
 {
-	"identifier": "amulet_team_editor",
+	"namespace": "amulet",
+	"identifier": "editor",
 	"version": "1.0.0",
 	"name": "Amulet Editor",
 	"depends": {
