+++ src/builtin_plugins/plugin/amulet/editor/window/_child/child_window_ui.ui
@@ -19,7 +19,7 @@
   <customwidget>
    <class>RecursiveSplitter</class>
    <extends>QFrame</extends>
-   <header>plugin.amulet_team_editor.window._tab_engine</header>
+   <header>plugin.amulet.editor.window._tab_engine</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
