+++ src/builtin_plugins/plugin/amulet/editor/window/_tab_engine.py
