+++ src/builtin_plugins/plugin/amulet_team_3d_viewer/_plugin.py
@@ -10,7 +10,7 @@ from amulet.app.plugin import PluginV1
 
 from plugin import tablericons
 from plugin.amulet_team_level import get_level
-from plugin.amulet_team_editor import (
+from plugin.amulet.editor import (
     register_widget,
     unregister_widget,
     register_layout,
